Moving some elements of LinearizedControlProblem to ControlProblem
